requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
Django==3.1.3
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
pyuploadcare==2.7.0
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
